Final touches
After running the code, you may get diffeent outputs at different run. It is because of the greedy algorithm, as it takes the local maxima or the minima, instead of the global maxima or minima.
I shall soon upload a more efficient algorithm, which shall optimize the problem in a much better way.